Entreega de proyecto completo

+++ database/migrations/2019_05_15_042713_create_plato_table.php
@@ -16,7 +16,7 @@ class CreatePlatoTable extends Migration
         Schema::create('plato', function (Blueprint $table) {
             $table->increments('Codigo');
             $table->char('Nombre',50);
-            $table->float('Valor');
+            $table->double('Valor');
             $table->timestamps();
         });
     }
